pyproject.toml
[project]
name = "bookbrain-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.128.0",
    "gunicorn>=24.1.1",
    "jinja2>=3.1.6",
    "numpy>=2.4.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.46",
    "stubs>=1.0.0",
    "uvicorn[standard]>=0.40.0",
]
